Ad Tracking: Facebook - track once per order, not per product
Before this change, each product on the order (the plan, the domain) were tracked was sent to FB separately, with its own parameter.

After this change, the order will be tracked only once, with the total cart price. The product_slug parameter will concatenate the products that are contained in the order.